replace /cataloge to separate route
