+++ app.js
@@ -11,6 +11,7 @@ var routes = require('./routes/index');
 var categories = require('./routes/categories');
 var products = require('./routes/products');
 var vendors = require('./routes/vendors');
+var catalog = require('./routes/catalog');
 
 var app = express();
 
@@ -35,6 +36,7 @@ app.use('/', routes);
 app.use('/categories', categories);
 app.use('/products', products);
 app.use('/vendors', vendors);
+app.use('/catalog', catalog);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
