+++ routes/index.js
@@ -30,29 +30,5 @@ router.get('/', function(req, res) {
     });
 });
 
-router.get('/catalog/', function(req, res) {
-
-    client.get(req.env.url + 'categories/', function (categories) {
-
-        var categoryArray = [];
-        for (var i = 0; i < categories.items.length; i++) {
-            var picName = categories.items[i].ico;
-            categoryArray[i] = {
-                name: categories.items[i].name,
-                id: categories.items[i]._id
-            };
-            if (picName !== undefined) {
-                var src = req.env.imUrl + picName[0] + '/' + picName[1] + '/' + picName[2] + '/' + picName;
-                categoryArray[i].src = src;
-            }
-        }
-
-        res.render('catalog', {
-            title: 'catalog',
-            categories: categoryArray
-        });
-    });
-});
-
 
 module.exports = router;
