+++ views/catalog.jade
@@ -1,7 +1,7 @@
 extends layout
 block content
   each category, i in categories
-    a(href='/categories/#{category._id}')
+    a(href='/categories/#{category.id}')
       p #{category.name}
       if category.src
         img(src='#{category.src}', width=200)
